+++ scripts/CoverEffect.js
@@ -84,7 +84,11 @@ export class CoverEffect {
    * Get the default data for this effect.
    * @returns {object}
    */
-  get defaultCoverObjectData() { return duplicate(this.constructor.defaultCoverObjectData.get(this.id)); }
+  get defaultCoverObjectData() {
+    const defaultData = this.constructor.defaultCoverObjectData.get(this.id);
+    if ( !defaultData ) return undefined;
+    return duplicate(defaultData);
+  }
 
   /**
    * Get the stored settings data for this effect.
@@ -93,11 +97,12 @@ export class CoverEffect {
 
   /**
    * Get the default document data for this effect.
-   * @returns {object}
+   * @returns {object|undefined}
    */
   get defaultDocumentData() {
-    const template = this.constructor.newCoverObjectData;
     const data = this.defaultCoverObjectData;
+    if ( !data ) return undefined;
+    const template = this.constructor.newCoverObjectData;
     const doc = foundry.utils.mergeObject(template, data.document, { inplace: false });
     foundry.utils.mergeObject(doc.flags[MODULE_ID], this.savedCoverRules, { inplace: true });
     doc.name = game.i18n.localize(data.name);
@@ -167,7 +172,26 @@ export class CoverEffect {
   get includeWalls() { return this.document.flags?.[MODULE_ID]?.[FLAGS.COVER_EFFECT.RULES.INCLUDE_WALLS]; }
 
   /** @type {boolean} */
-  get includeTokens() { return this.document.flags?.[MODULE_ID]?.[FLAGS.COVER_EFFECT.RULES.INCLUDE_TOKENS]; }
+  get liveTokensBlock() { return this.document.flags?.[MODULE_ID]?.[FLAGS.COVER_EFFECT.RULES.LIVE_TOKENS_BLOCK]; }
+
+  /** @type {boolean} */
+  get deadTokensBlock() { return this.document.flags?.[MODULE_ID]?.[FLAGS.COVER_EFFECT.RULES.DEAD_TOKENS_BLOCK]; }
+
+  /** @type {boolean} */
+  get proneTokensBlock() { return this.document.flags?.[MODULE_ID]?.[FLAGS.COVER_EFFECT.RULES.PRONE_TOKENS_BLOCK]; }
+
+  /** @type {boolean} */
+  get includeTokens() { return this.liveTokensBlock || this.deadTokensBlock; }
+
+  /** @type {AlternativeLOSConfig} */
+  get calcConfig() {
+    return {
+      deadTokensBlock: this.deadTokensBlock,
+      liveTokensBlock: this.liveTokensBlock,
+      proneTokensBlock: this.proneTokensBlock,
+      wallsBlock: this.includeWalls
+    };
+  }
 
 
   // ----- NOTE: Calculation methods ----- //
@@ -180,7 +204,9 @@ export class CoverEffect {
    */
   percentCover(attackingToken, targetToken) {
     const { includeWalls, includeTokens } = this;
-    return attackingToken.tokencover.coverCalculator.percentCover(targetToken, { includeWalls, includeTokens });
+    const calc = attackingToken.tokencover.coverCalculator;
+    calc.updateConfiguration(this.calcConfig);
+    return calc.percentCover(targetToken, { includeWalls, includeTokens });
   }
 
   /**
@@ -442,7 +468,9 @@ export class CoverEffect {
           [FLAGS.COVER_EFFECT.RULES.PRIORITY]: 0,
           [FLAGS.COVER_EFFECT.RULES.OVERLAPS]: false,
           [FLAGS.COVER_EFFECT.RULES.INCLUDE_WALLS]: true,
-          [FLAGS.COVER_EFFECT.RULES.INCLUDE_TOKENS]: false
+          [FLAGS.COVER_EFFECT.RULES.LIVE_TOKENS_BLOCK]: false,
+          [FLAGS.COVER_EFFECT.RULES.DEAD_TOKENS_BLOCK]: false,
+          [FLAGS.COVER_EFFECT.RULES.PRONE_TOKENS_BLOCK]: false,
         }
       }
     }
@@ -610,7 +638,7 @@ export class CoverEffect {
    */
   static async initialize() {
     let storedIds = this.storedCoverObjectIds;
-    if ( !storedIds.size ) storedIds = this.defaultCoverObjectData.keys();
+    if ( !storedIds.size ) storedIds = this.defaultCoverObjectData?.keys();
     for ( const id of storedIds ) await this.create(id);
     this.transitionDocuments();
   }